// apps/gatehouse-service-web/src/.generated/api/model/apiErrorResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { ApiErrorResponseDtoData } from './apiErrorResponseDtoData';
import type { ApiErrorResponseDtoMeta } from './apiErrorResponseDtoMeta';
import type { ApiValidationErrorDetailDto } from './apiValidationErrorDetailDto';

export interface ApiErrorResponseDto {
  /** 성공 여부 (에러시 false) */
  success: boolean;
  /** HTTP 상태 코드 */
  statusCode: number;
  /** 요청 경로 */
  path: string;
  /** 요청 ID */
  requestId: string;
  /** 응답 생성 일시 */
  timestamp: string;
  /** 에러 메시지 */
  message?: string;
  /**
     * 응답 데이터
     * @nullable
     */
  data: ApiErrorResponseDtoData;
  /** 메타 데이터 */
  meta?: ApiErrorResponseDtoMeta;
  /** 에러 코드 */
  errorCode: string;
  /**
     * 에러 세부 정보
     * @nullable
     */
  details?: ApiValidationErrorDetailDto[] | null;
}